// software-engineering/space-api-web/src/app/models/Space.ts
export class Space {
    copyright: string;
    date: string;
    explanation: string;
    hdurl: string;
    media_type: string;
    service_version: string;
    title: string;
    url: string;

    constructor () {
        this.copyright = '',
        this.date = '',
        this.explanation = '',
        this.hdurl = '',
        this.media_type = '',
        this.service_version = '',
        this.title = '',
        this.url = ''
    }
}